requirements.txt
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.0
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.6
comm==0.1.2
command-not-found==0.3
contourpy==1.0.6
cryptography==3.4.8
cssselect==1.1.0
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.5
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==3.2.12
dnspython==2.3.0
duplicity==0.8.21
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.6.0
fonttools==4.38.0
fqdn==1.5.1
future==0.18.2
greenlet==2.0.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
install==1.3.5
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mongo==0.2.0
monotonic==1.6
more-itertools==8.10.0
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
oauthlib==3.2.0
olefile==0.46
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.6.2
pluggy==1.0.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.5
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.3.3
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
pytest==7.2.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-json-logger==2.0.4
pytz==2022.7.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
reportlab==3.6.8
requests==2.25.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.1
scipy==1.10.0
scour==0.38.2
seaborn==0.12.2
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
sqlparse==0.4.2
stack-data==0.6.2
statsmodels==0.13.5
systemd-python==234
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.8.1
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
xdg==5
xkit==0.0.0
zipp==1.0.0
attrs==22.2.0
autopep8==2.0.1
colorama==0.4.6
dnspython==2.3.0
exceptiongroup==1.1.0
iniconfig==2.0.0
mongo==0.2.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.10.0
pymongo==4.3.3
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
tomli==2.0.1
